Move type definitions from cgi.h to separate file
This way consumers can include the new cgi_types.h only, and do not need
to include the whole cgi.h.

Bump version, so you can depend on that new header file.
