+++ CMakeLists.txt
@@ -8,7 +8,7 @@
 cmake_minimum_required(VERSION 3.1)
 
 project(cgi
-	VERSION 1.2.0
+	VERSION 1.2.1
 	LANGUAGES C
 )
 message(STATUS "cgi_VERSION: ${cgi_VERSION}")
