+++ test/test.c
@@ -16,6 +16,7 @@
 #include "cgi_test.h"
 
 #include "libcgi/cgi.h"
+#include "libcgi/cgi_types.h"
 #include "libcgi/config.h"
 
 extern formvars *
